Optimizer object, initial commit
